自用修改

+++ clash/hy2.sh
@@ -24,9 +24,13 @@ check_log_file() {
 check_log_file "$LOG_FILE"
 add_log() {
     local new_content=$1
-    existing_content=$(cat "$LOG_FILE")
-    combined_content="$new_content\n$existing_content"
-    echo -e "$combined_content" > "$LOG_FILE"
+    if [ -f "$LOG_FILE" ]; then
+        existing_content=$(cat "$LOG_FILE")
+        combined_content="$new_content\n$existing_content"
+        echo -e "$combined_content" > "$LOG_FILE"
+    else
+        echo -e "$new_content" > "$LOG_FILE"
+    fi
 }
 
 check_ip() {
@@ -44,49 +48,41 @@ check_ip() {
 # 检查是否有 "hysteria2" 的进程在运行
 process_status=$(pgrep -f "config.yaml" >/dev/null 2>&1; echo $?)
 if [ $process_status -eq 0 ]; then
-    echo "hysteria2 进程正在运行..."
-    add_log "${DATE_FORMAT} - hysteria2 进程正在运行..."
+    echo "hy2 进程正在运行..."
+    add_log "${DATE_FORMAT} - hy2 is running..."
     if [ -f "$HTML_DIR/cg.json" ]; then
         C_IP=$(grep '"ip"' "$HTML_DIR/cg.json" | sed 's/.*"ip": "\(.*\)",/\1/')
         if check_ip "$C_IP"; then
             exit 0
         else
-            add_log "${DATE_FORMAT} - hy2进程不存在，准备重新安装 hy2..."
+            add_log "${DATE_FORMAT} - hy2 not exist，start reinstall hy2..."
         fi
     fi
 else
-    add_log "${DATE_FORMAT} - hy2进程不存在，准备重新安装 hy2..."
+    add_log "${DATE_FORMAT} - hy2 not exist，start reinstall hy2..."
 fi
 
 [[ "$HOSTNAME" == "s1.ct8.pl" ]] && DOMAINS=("s1.ct8.pl" "cache.ct8.pl" "web.ct8.pl" "panel.ct8.pl") || DOMAINS=("s${NUM}.serv00.com" "cache${NUM}.serv00.com" "web${NUM}.serv00.com" "panel${NUM}.serv00.com")
 
 ip=$(curl -s --max-time 1.5 ipv4.ip.sb)
+if [ -n "$ip" ]; then
+    if! check_ip "$ip"; then
+        $ip= ""
+    fi
+fi
 if [ -z "$ip" ]; then
     for domain in "${DOMAINS[@]}"; do
         echo "检查域名是否可用 $domain"
         if check_ip "$domain"; then
             echo "域名 $domain 可用"
-            ip="$domain"
+            ip = "$domain"
             break  # 域名可用，跳出循环
         else
             echo "域名 $domain 不可用"
         fi
     done
-else 
-    if ! check_ip "$ip"; then
-        ip=""
-        for domain in "${DOMAINS[@]}"; do
-            echo "检查域名是否可用 $domain"
-            if check_ip "$domain"; then
-                echo "域名 $domain 可用"
-                ip="$domain"
-                break  # 域名可用，跳出循环
-            else
-                echo "域名 $domain 不可用"
-            fi
-        done
-    fi
 fi
+
 if [ -z "$ip" ]; then
     HOST_IP=""
 elif [[ "$ip" =~ ^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$ ]]; then
@@ -105,7 +101,7 @@ cat <<EOF > $HTML_DIR/cg.json
 EOF
 # 判断 HOST_IP 是否为空
 if [ -z "$HOST_IP" ]; then
-  add_log "${DATE_FORMAT} - 找不到可用IP..."
+  add_log "${DATE_FORMAT} - unable to find available ip."
   echo "找不到可用IP，开始停止进程..."
   pkill -u $USERNAME
   exit 0
@@ -255,5 +251,7 @@ EOF
 rm -rf config.yaml fake_useragent_0.2.0.json
 echo -e "\n\e[1;32mRuning done!\033[0m"
 echo -e "\e[1;35m原脚本地址：https://github.com/eooce/scripts\e[0m"
+
+add_log "${DATE_FORMAT} - hy2 install success."
 
 exit 0
